Rewrite + Files Added
Rewrote "database_manager.py" as a class for organization. Anything with "pass" in the function body still needs to be implemented, excluding the __init__() statement.

Renamed the cars database from 'car_Rental.db' to 'cars.db'

Added new files "database_logger.py" (To be implemented) and "database_duplicator.py" (To be implemented).